Clean: Relocate STEX WatershedSegmentation
